gpu: nvgpu: volt: fix MISRA 10.3 violations
Fix MISRA Rule 10.3 violations in common/pmu/volt for assigning
objects of different size or essential type.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/2027657
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/include/nvgpu/pmuif/gpmuifvolt.h
@@ -1,6 +1,6 @@
 /*
-* Copyright (c) 2016-2018, NVIDIA CORPORATION.  All rights reserved.
-*
+ * Copyright (c) 2016-2019, NVIDIA CORPORATION.  All rights reserved.
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
  * to deal in the Software without restriction, including without limitation
@@ -267,7 +267,7 @@ struct nv_pmu_volt_cmd_rpc {
 };
 
 #define NV_PMU_VOLT_CMD_RPC_ALLOC_OFFSET	\
-	offsetof(struct nv_pmu_volt_cmd_rpc, request)
+	(u32)offsetof(struct nv_pmu_volt_cmd_rpc, request)
 
 struct nv_pmu_volt_cmd {
 	union {
@@ -308,7 +308,7 @@ struct nv_pmu_volt_msg_rpc {
 };
 
 #define NV_PMU_VOLT_MSG_RPC_ALLOC_OFFSET	\
-		offsetof(struct nv_pmu_volt_msg_rpc, response)
+		(u32)offsetof(struct nv_pmu_volt_msg_rpc, response)
 
 struct nv_pmu_volt_msg {
 	union {
